pyproject.toml
[project]
name = "workshop_airflow"
version = "0.1.0"
description = ""
authors = [
    {name = "sch-paulo",email = "pschmitzjr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "streamlit (>=1.42.0,<2.0.0)",
    "apache-airflow (>=2.10.5,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)"
]
